requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.3
djangorestframework==3.14.0
gcloud==0.18.3
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.6
jwcrypto==1.5.6
oauth2client==4.1.3
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pyparsing==3.1.2
Pyrebase4==4.7.1
python-jwt==4.1.0
pytz==2024.1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==1.26.18
